[Move `@kbn/config-schema` to server] `maps`
