+++ x-pack/plugins/maps/public/index.ts
@@ -8,7 +8,7 @@
 import { PluginInitializer } from '@kbn/core/public';
 import { PluginInitializerContext } from '@kbn/core/public';
 import { MapsPlugin, MapsPluginSetup, MapsPluginStart } from './plugin';
-import { MapsXPackConfig } from '../config';
+import type { MapsXPackConfig } from '../server/config';
 
 export const plugin: PluginInitializer<MapsPluginSetup, MapsPluginStart> = (
   initContext: PluginInitializerContext<MapsXPackConfig>
