+++ x-pack/plugins/maps/public/kibana_services.ts
@@ -9,7 +9,7 @@ import type { CoreStart } from '@kbn/core/public';
 import type { EMSSettings } from '@kbn/maps-ems-plugin/common/ems_settings';
 import { MapsEmsPluginPublicStart } from '@kbn/maps-ems-plugin/public';
 import { BehaviorSubject } from 'rxjs';
-import type { MapsConfigType } from '../config';
+import type { MapsConfigType } from '../server/config';
 import type { MapsPluginStartDependencies } from './plugin';
 
 const servicesReady$ = new BehaviorSubject(false);
