+++ x-pack/plugins/maps/server/index.ts
@@ -7,7 +7,7 @@
 
 import { PluginInitializerContext } from '@kbn/core/server';
 import { PluginConfigDescriptor } from '@kbn/core/server';
-import { configSchema, MapsXPackConfig } from '../config';
+import { configSchema, MapsXPackConfig } from './config';
 
 export const config: PluginConfigDescriptor<MapsXPackConfig> = {
   // exposeToBrowser specifies kibana.yml settings to expose to the browser
